+++ CrashReporting/Sources/Domain/Services/CrashReportingServiceAdapter.swift
@@ -11,9 +11,25 @@ import Logger
 class CrashReportingServiceAdapter: CrashReportingService {
 
     var apiService: APIService
+    var logger: LogProtocol
+    var botLogHandler: (() -> String?)?
 
-    init(_ apiService: APIService) {
+    init(_ apiService: APIService, logger: LogProtocol = Log.shared) {
         self.apiService = apiService
+        self.logger = logger
+        self.apiService.onEventHandler = { [weak self] () -> Logs in
+            var appLog: String?
+            if let logUrls = logger.fileUrls.first {
+                do {
+                    let data = try Data(contentsOf: logUrls)
+                    appLog = String(data: data, encoding: .utf8)
+                } catch {
+                    logger.optional(error, nil)
+                }
+            }
+            let botLog = self?.botLogHandler?()
+            return Logs(appLog: appLog, botLog: botLog)
+        }
     }
 
     func identify(identity: Identity) {
@@ -25,21 +41,12 @@ class CrashReportingServiceAdapter: CrashReportingService {
     }
 
     func record(_ message: String) {
-        Log.debug(message)
+        logger.debug(message)
         apiService.record(message)
     }
 
-    func report(error: Error, metadata: [AnyHashable: Any]? = nil, botLog: String?) {
-        Log.error(error.localizedDescription)
-        var appLog: String?
-        if let logUrls = Log.fileUrls.first {
-            do {
-                let data = try Data(contentsOf: logUrls)
-                appLog = String(data: data, encoding: .utf8)
-            } catch {
-                Log.optional(error)
-            }
-        }
-        apiService.report(error: error, metadata: metadata, appLog: appLog, botLog: botLog)
+    func report(error: Error, metadata: [AnyHashable: Any]? = nil) {
+        logger.error(error.localizedDescription)
+        apiService.report(error: error, metadata: metadata)
     }
 }