+++ CrashReporting/Tests/Mocks/APIServiceMock.swift
@@ -9,11 +9,12 @@ import Foundation
 @testable import CrashReporting
 
 class APIServiceMock: APIService {
-
+    var onEventHandler: (() -> Logs)?
     var identified = false
     var crashed = false
     var forgot = false
     var recorded = false
+    var lastAttachedLogs: Logs?
 
     func identify(identity: Identity) {
         identified = true
@@ -27,7 +28,8 @@ class APIServiceMock: APIService {
         recorded = true
     }
 
-    func report(error: Error, metadata: [AnyHashable: Any]?, appLog: String?, botLog: String?) {
+    func report(error: Error, metadata: [AnyHashable: Any]?) {
         crashed = true
+        lastAttachedLogs = onEventHandler?()
     }
 }