+++ Logger/Sources/Delivery/Log.swift
@@ -14,14 +14,7 @@ import Foundation
 /// - ERROR or UNEXPECTED: A handleable error condition.
 /// - INFO:  Generic (useful) information about system operation.
 /// - DEBUG: Low-level information for developers.
-public class Log {
-
-    public enum Reason: String {
-        case apiError
-        case botError
-        case missingValue
-        case incorrectValue
-    }
+public class Log: LogProtocol {
 
     public static let shared = Log()
 
@@ -46,6 +39,10 @@ public class Log {
 
     public func debug(_ string: String) {
         service.debug(string)
+    }
+
+    public func error(_ string: String) {
+        service.unexpected(string, nil)
     }
 
     public func unexpected(_ reason: Reason, _ detail: String?) {
@@ -110,6 +107,6 @@ public extension Log {
 
     /// Log a ERROR message
     static func error(_ message: String) {
-        shared.service.unexpected(message, nil)
+        shared.error(message)
     }
 }